feat: cargo xtask themelint <theme>
