+++ xtask/Cargo.toml
@@ -8,4 +8,5 @@ edition = "2021"
 [dependencies]
 helix-term = { version = "0.6", path = "../helix-term" }
 helix-core = { version = "0.6", path = "../helix-core" }
+helix-view = { version = "0.6", path = "../helix-view" }
 toml = "0.5"
